requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
deepl==1.18.0
dill==0.3.9
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.1
executing==2.1.0
filelock==3.13.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.10
imageio==2.35.1
ipython==8.28.0
ipywidgets==8.1.5
jax==0.4.34
jax_dataclasses==1.6.1
jaxlib==0.4.34
jaxlie==1.4.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyterlab_widgets==3.0.13
k3d==2.16.1
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
lightning==2.4.0
lightning-utilities==0.11.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
networkx==3.3
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
openai==1.32.0
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plyfile==1.1
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytorch-lightning==2.4.0
PyYAML==6.0.2
regex==2024.5.15
requests==2.32.3
rich==13.9.2
roma==1.5.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.1
selenium==4.21.0
sentence-transformers==3.0.1
sentry-sdk==2.15.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
torchmetrics==1.4.2
ninja==1.11.1.1
healpy==1.17.3
piqa==1.3.2
lpips==0.1.4
opencv-python==4.10.0.84
pandas==2.2.3
seaborn==0.13.2
plotly==5.24.1
tqdm==4.66.4
traitlets==5.14.3
traittypes==0.2.1
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
triton==3.0.0
typing_extensions==4.12.1
tyro==0.8.11
urllib3==2.2.3
wandb==0.18.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
wsproto==1.2.0
yarl==1.13.1